+++ barapost/src/binning_modules/single_thread_FAST5_utwfunc.py
@@ -9,7 +9,7 @@ from shelve import open as open_shelve
 
 from src.fmt_readID import fmt_read_id
 from src.platform import platf_depend_exit
-from src.sorter_modules.fast5 import fast5_readids
+from src.binning_modules.fast5 import fast5_readids
 from src.printlog import printl, printn, getwt, err_fmt
 
 index_name = "fast5_to_tsvtaxann_idx"