+++ barapost/src/binning_modules/single_thread_QA.py
@@ -1,47 +1,47 @@
 # -*- coding: utf-8 -*-
-# Module defines functions necessary for sorting FASTA and FASTQ files in single thread.
+# Module defines functions necessary for binning FASTA and FASTQ files in single thread.
 
 import os
 import sys
 
-from src.sorter_modules.sorter_spec import *
+from src.binning_modules.binning_spec import *
 
 from src.fmt_readID import fmt_read_id
 from src.platform import platf_depend_exit
 from src.printlog import printl, printn, getwt, err_fmt
 from src.filesystem import get_curr_res_dpath, is_fastq
 
-from src.sorter_modules.fastq_records import fastq_records
-from src.sorter_modules.fasta_records import fasta_records
+from src.binning_modules.fastq_records import fastq_records
+from src.binning_modules.fasta_records import fasta_records
 
-from src.sorter_modules.filters import get_QL_filter, get_QL_trash_fpath
-from src.sorter_modules.filters import get_align_filter, get_align_trash_fpath
+from src.binning_modules.filters import get_QL_filter, get_QL_trash_fpath
+from src.binning_modules.filters import get_align_filter, get_align_trash_fpath
 
 
-def write_fastq_record(sorted_file, fastq_record):
+def write_fastq_record(binned_file, fastq_record):
     """
-    :param sorted_path: path to file, in which data from fastq_record is oing to be written;
-    :type sorted_path: str;
+    :param binned_file: file instance, in which data from fastq_record is oing to be written;
+    :type binned_file: _oi.TextIOWrapper;
     :param fastq_record: dict of 4 elements. Elements are four corresponding lines of FASTQ;
     :type fastq_record: dict<str: str>;
     """
 
-    sorted_file.write(fastq_record["seq_id"]+'\n')
-    sorted_file.write(fastq_record["seq"]+'\n')
-    sorted_file.write(fastq_record["opt_id"]+'\n')
-    sorted_file.write(fastq_record["qual_line"]+'\n')
+    binned_file.write(fastq_record["seq_id"]+'\n')
+    binned_file.write(fastq_record["seq"]+'\n')
+    binned_file.write(fastq_record["opt_id"]+'\n')
+    binned_file.write(fastq_record["qual_line"]+'\n')
 # end def write_fastq_record
 
 
-def write_fasta_record(sorted_file, fasta_record):
+def write_fasta_record(binned_file, fasta_record):
     """
-    :param sorted_file: file, which data from fasta_record is written in
-    :type sorted_file: _io.TextIOWrapper
+    :param binned_file: file, which data from fasta_record is written in
+    :type binned_file: _io.TextIOWrapper
     :param fasta_record: dict of 2 elements. Elements are four corresponding lines of FASTA
     :type fasta_record: dict<str: str>
     """
-    sorted_file.write(fasta_record["seq_id"]+'\n')
-    sorted_file.write(fasta_record["seq"]+'\n')
+    binned_file.write(fasta_record["seq_id"]+'\n')
+    binned_file.write(fasta_record["seq"]+'\n')
 
 # end def write_fasta_record
 
@@ -59,16 +59,16 @@ def update_file_dict(srt_file_dict, new_fpath, logfile_path):
 # end def update_file_dict
 
 
-def sort_fastqa_file(fq_fa_path, tax_annot_res_dir, sens,
+def bin_fastqa_file(fq_fa_path, tax_annot_res_dir, sens,
         min_qual, min_qlen, min_pident, min_coverage, logfile_path):
     """
-    Function for single-thread sorting FASTQ and FASTA files.
+    Function for single-thread binning FASTQ and FASTA files.
 
     :param fq_fa_path: path to FASTQ (of FASTA) file meant to be processed;
     :type fq_fa_path: str;
     :param tax_annot_res_dir: path to directory containing taxonomic annotation;
     :type tax_annot_res_dir: str;
-    :param sens: sorting sensitivity;
+    :param sens: binning sensitivity;
     :type sens: str;
     :param min_qual: threshold for quality filter;
     :type min_qual: float;
@@ -133,36 +133,36 @@ This TSV file: '{}'""".format(read_name, tsv_res_fpath)))
             if QL_trash_fpath not in srt_file_dict.keys():
                 srt_file_dict = update_file_dict(srt_file_dict, QL_trash_fpath, logfile_path)
             # end if
-            write_fun(srt_file_dict[QL_trash_fpath], fastq_rec) # write current read to sorted file
+            write_fun(srt_file_dict[QL_trash_fpath], fastq_rec) # write current read to binned file
             QL_seqs_fail += 1
         elif not align_filter(vals_to_filter):
             # Place this sequence to align_trash file
             if align_trash_fpath not in srt_file_dict.keys():
                 srt_file_dict = update_file_dict(srt_file_dict, align_trash_fpath, logfile_path)
             # end if
-            write_fun(srt_file_dict[align_trash_fpath], fastq_rec) # write current read to sorted file
+            write_fun(srt_file_dict[align_trash_fpath], fastq_rec) # write current read to binned file
             align_seqs_fail += 1
         else:
             for hit_name in hit_names.split("&&"): # there can be multiple hits for single query sequence
                 # Get name of result FASTQ file to write this read in
-                sorted_file_path = os.path.join(outdir_path, "{}.fast{}".format(hit_name,
+                binned_file_path = os.path.join(outdir_path, "{}.fast{}".format(hit_name,
                     'q' if is_fastq(fq_fa_path) else 'a'))
-                if sorted_file_path not in srt_file_dict.keys():
-                    srt_file_dict = update_file_dict(srt_file_dict, sorted_file_path, logfile_path)
+                if binned_file_path not in srt_file_dict.keys():
+                    srt_file_dict = update_file_dict(srt_file_dict, binned_file_path, logfile_path)
                 # end if
-                write_fun(srt_file_dict[sorted_file_path], fastq_rec) # write current read to sorted file
+                write_fun(srt_file_dict[binned_file_path], fastq_rec) # write current read to binned file
             # end for
             seqs_pass += 1
         # end if
     # end for
 
-    # Close all sorted files
+    # Close all binned files
     for file_obj in srt_file_dict.values():
         file_obj.close()
     # end for
 
-    printl(logfile_path, "\r{} - File '{}' is sorted.".format(getwt(), os.path.basename(fq_fa_path)))
+    printl(logfile_path, "\r{} - File '{}' is binned.".format(getwt(), os.path.basename(fq_fa_path)))
     printn(" Working...")
 
     return (seqs_pass, QL_seqs_fail, align_seqs_fail)
-# end def sort_fastqa_file
+# end def bin_fastqa_file