+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # "Barapost" changelog
 
+## 2020-06-12 edition
+
+- README updated;
+- minor renaming changes in source files;
+
+## Version changes:
+
+- barapost-prober: `1.21.b --> 1.21.c`;
+- barapost-local `3.14.f --> 3.14.g`;
+- barapost-binning: `4.5.f --> 4.5.g`;
+
 ## 2020-06-11
 
 All scripts are renamed.